Improved mapping
The API URL for title list generation can be configured over the
config.properties file and will be loaded on startup.

All (available) titles will now be considered for the mapping ID. The
titles will be looked up and the most occurring element will be
selected.

Bad Requests will be returned with null instead of the String to prevent
JSON Parse exception.

The language will be considered for requesting the information of
TheTVDB and TheMovieDB.

The API URL for TheTVDB and TheMovieDB will now be encoded into ASCII
instead of the parameters just be encoded, this should increase the
number of resulting mappings.